First draft of the JF sections from PSI.
Suggest mentioning Alpaka in the title.
First draft of the abstract, sections 1 and 2 from Sophie, Martin, Carlos and Aldo.
Introduction will need merging with HZDR paragraph.
Few references added, and a few renamed for clarity.
We are not sure how long the final article should be. This work could be too detailed, in which case we can cut down.